Rename based on `flutter_lints`
