+++ lib/src/config/global_config.dart
